// Api/src/types/permissions.ts
export enum PermissionsEnum {
  CREATE_PORTAL_USERS = 'Create Portal Users',

  CREATE_HUB_ADMIN = 'Create Hub Admin',

  CREATE_DFSP_ADMIN = 'Create DFSP Admin',
  CREATE_DFSP_OPERATOR = 'Create DFSP Operator',
  CREATE_DFSP_AUDITOR = 'Create DFSP Auditor',

  VIEW_ROLES = 'View Roles',

  VIEW_DFSPS = 'View DFSPs',
  CREATE_DFSPS = 'Create DFSPs',
  EDIT_DFSPS = 'Edit DFSPs',
  DELETE_DFSPS = 'Delete DFSPs',

  VIEW_PORTAL_USERS = 'View Portal Users',
  EDIT_PORTAL_USERS_STATUS = 'Edit Portal Users Status',

  VIEW_MERCHANTS = 'View Merchants',
  CREATE_MERCHANTS = 'Create Merchants',
  EDIT_MERCHANTS = 'Edit Merchants',

  EXPORT_MERCHANTS = 'Export Merchants',

  APPROVE_MERCHANTS = 'Approve Merchants',
  REJECT_MERCHANTS = 'Reject Merchants',
  REVERT_MERCHANTS = 'Revert Merchants',

  // Not used
  DELETE_MERCHANTS = 'Delete Merchants',

  // Not used
  CREATE_ROLES = 'Create Roles',
  EDIT_ROLES = 'Edit Roles',

  // Not used
  EDIT_PORTAL_USERS = 'Edit Portal Users',
  DELETE_PORTAL_USERS = 'Delete Portal Users',

  VIEW_AUDIT_LOGS = 'View Audit Logs',

  EDIT_SERVER_LOG_LEVEL = 'Edit Server Log Level',

  VIEW_TRANSACTIONS = 'View Transactions',


  
  VIEW_RECONCILIATIONS = 'View Reconcilitions',
  CLOSE_RECONCILIATIONS = 'Close Reconcilitions',
  CHANGE_RECON_MISMATCH_STATUS = 'Change Reconcilition mismatch status'
}